New classes: InstanceParser and CommentsInserter
Added the class 'InstanceParser' as a thin wrapper around ASTParser with a
few convenience methods for parsing CompilationUnits, Blocks,
Statements, etc.

Added the class 'CommentsInsert' to encapsulat code for assigning comments
to nodes of the AST.

Enabled the COMMON_TOKEN_ACTION=true option in the java_1_8.jj file.

Implemented a CommonTokenAction(Token token) method in the
TOKEN_MGR_DECLS that collects encountered tokens in a list.

Added a getter for this list of tokens in the ASTParser.